pyproject.toml
[tool.poetry]
name = "learning-large-language-models"
version = "0.1.0"
description = ""
authors = ["alexjenkins-86400 <alex.jenkins@86400.com.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11 <3.12"
openai = "^1.6.1"
pandas = "^2.1.4"
llama-index = "^0.9.22"
python-dotenv = "^1.0.0"
unstructured = "^0.11.6"
chromadb = "^0.4.21"
langchain-experimental = "^0.0.47"
beautifulsoup4 = "^4.12.2"
selenium = "^4.16.0"
fastapi = "^0.108.0"
lxml = "4.9.2"
torch = "^2.1.2"
transformers = "^4.36.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
